backend/requirements.txt
# FastAPI and ASGI
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg>=0.30.0          # PostgreSQL async driver
alembic==1.12.1          # Database migrations

# Data validation and serialization
pydantic>=2.10.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# SSO Integration
python3-saml==1.16.0       # SAML 2.0 support
authlib==1.3.0             # OAuth 2.0 and OIDC support
cryptography>=41.0.0       # Enhanced crypto for SSO
PyJWT[crypto]==2.8.0       # JWT token handling for OIDC

# Background tasks and caching
celery==5.3.4
redis==5.0.1

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security scanning
bandit==1.7.5

# Logging and monitoring
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0

# Environment and configuration
python-dotenv==1.0.0

# JIRA and Confluence integration
atlassian-python-api==3.41.3

# Timezone handling
pytz==2023.3

# Data processing
pandas>=2.2.0           # For analytics and reporting
numpy>=1.26.0           # Mathematical operations